packages/pulse-aws/pyproject.toml
[project]
name = "pulse-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Erwin Kuhn", email = "erwin.kuhn@protonmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.160.0,<3.0",
    "constructs>=10.0.0,<11.0.0",
    "boto3>=1.35.0,<2.0.0",
    "botocore>=1.35.0,<2.0.0",
    "httpx>=0.27.0,<1.0.0",
]

[tool.uv]
package = true

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
